third_party/WebKit/LayoutTests/editing/selection/caret-in-textarea-auto.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<textarea id="textarea" dir="auto" style="font-size: 20px; width: 20ex; border: solid thin black; padding: 10px;">&#x05d0;!</textarea>
<script>
description('Test caret position in a textarea with direction auto and having RTL text. When entering a LTR text caret position is changed.');
var textarea = document.getElementById('textarea');
textarea.focus();
var caretRect = textInputController.firstRectForCharacterRange(0, 0);

window.testRunner.execCommand('MoveToEndOfLine');
document.execCommand('InsertLineBreak');
document.execCommand('InsertText', false, 'hell');

var caretRect1 = textInputController.firstRectForCharacterRange(1, 0);
shouldNotBe('caretRect[0]', 'caretRect1[0]');
</script>
