third_party/WebKit/LayoutTests/editing/selection/caret-in-textarea-auto-expected.txt

Test caret position in a textarea with direction auto and having RTL text. When entering a LTR text caret position is changed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS caretRect[0] is not caretRect1[0]
PASS successfullyParsed is true

TEST COMPLETE
